requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1

# additional requirements
numpy
scikit-learn>=1.3.0
notebook
matplotlib
seaborn
imageio
pandas
backpack-for-pytorch
tqdm